SystemTitanic
Work for collge for the subject of Algorithm and Computer Programming 2, it was proposed to make a software to manage crew and passengers of the Titanic accident. I work with the intention of learning how to use the file modification functions. The teacher passed the file to modify with the informations and from that the code was being made.